ai-logging-agent/pyproject.toml
[project]
name = "ai-logging-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
"openai-agents[litellm]>=0.0.19",
"python-dotenv",
"requests"
]
